requirements_mambajscc.txt
torch==2.6.0
mamba-ssm==2.2.4
timm==1.0.20
tensorboardX==2.6.2.2
einops==0.8.1
transformers==4.50.0
tokenizers==0.21.4
Pillow==11.1.0
requests==2.32.5
torchvision
numpy>=2,<2.3
matplotlib
contourpy>=1.0.1
cycler>=0.10
fonttools>=4.22.0
kiwisolver>=1.3.1
pyparsing>=3
python-dateutil>=2.7
protobuf>=3.20
scipy
pytorch-msssim
opencv-python
